Add email protection flag plumbing and tests; don't call generate bundle when making an intermediate CSR since everything is now ignored

+++ builtin/logical/pki/path_config_ca.go
@@ -273,7 +273,7 @@ func (b *backend) pathCAGenerateRoot(
 			"serial_number": cb.SerialNumber,
 			"certificate":   cb.Certificate,
 			"issuing_ca":    cb.IssuingCA,
-			"expiration":    int(parsedBundle.Certificate.NotAfter.Unix()),
+			"expiration":    int64(parsedBundle.Certificate.NotAfter.Unix()),
 		},
 	}
